refactor: 모달 경로 수정
